refactor: move the process of the webpack config integration to other module
